pyproject.toml
[tool.poetry]
name = "sam-aging"
version = "0.1.0"
description = ""
authors = ["whatphilipcodes <gerdes.philip@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
